move user iterator common code to mod.rs
