+++ src/storage/src/hummock/iterator/mod.rs
@@ -14,7 +14,7 @@
 
 use std::future::Future;
 use std::marker::PhantomData;
-use std::ops::{Deref, DerefMut};
+use std::ops::{Bound, Deref, DerefMut};
 
 use super::{HummockResult, HummockValue};
 
@@ -36,7 +36,10 @@ pub use forward_user::*;
 pub use merge_inner::{OrderedMergeIteratorInner, UnorderedMergeIteratorInner};
 
 use crate::hummock::iterator::HummockIteratorUnion::{First, Fourth, Second, Third};
-use crate::hummock::SstableIterator;
+use crate::hummock::local_version::PinnedVersion;
+use crate::hummock::shared_buffer::shared_buffer_batch::SharedBufferBatchIterator;
+use crate::hummock::shared_buffer::SharedBufferIteratorType;
+use crate::hummock::{BackwardSstableIterator, SstableIterator, SstableIteratorType};
 
 #[cfg(any(test, feature = "test"))]
 pub mod test_utils;
@@ -336,3 +339,98 @@ impl HummockIteratorDirection for Backward {
 
 pub type MultiSstIterator =
     UnorderedMergeIteratorInner<HummockIteratorUnion<Forward, ConcatIterator, SstableIterator>>;
+
+#[allow(type_alias_bounds)]
+pub type UserIteratorPayloadType<
+    D: HummockIteratorDirection,
+    I: SstableIteratorType<Direction = D>,
+> = HummockIteratorUnion<
+    D,
+    SharedBufferBatchIterator<D>,
+    SharedBufferIteratorType<D, I>,
+    ConcatIteratorInner<I>,
+    I,
+>;
+
+pub type ForwardUserIteratorType =
+    UnorderedMergeIteratorInner<UserIteratorPayloadType<Forward, SstableIterator>>;
+pub type BackwardUserIteratorType =
+    UnorderedMergeIteratorInner<UserIteratorPayloadType<Backward, BackwardSstableIterator>>;
+
+pub enum DirectedUserIterator {
+    Forward(UserIterator<ForwardUserIteratorType>),
+    Backward(BackwardUserIterator<BackwardUserIteratorType>),
+}
+
+pub trait DirectedUserIteratorBuilder {
+    type Direction: HummockIteratorDirection;
+    type SstableIteratorType: SstableIteratorType<Direction = Self::Direction>;
+    /// Initialize an `DirectedUserIterator`.
+    /// The `key_range` should be from smaller key to larger key.
+    fn create(
+        iterator_iter: impl IntoIterator<
+            Item = UserIteratorPayloadType<Self::Direction, Self::SstableIteratorType>,
+        >,
+        key_range: (Bound<Vec<u8>>, Bound<Vec<u8>>),
+        read_epoch: u64,
+        min_epoch: u64,
+        version: Option<PinnedVersion>,
+    ) -> DirectedUserIterator;
+}
+
+impl DirectedUserIterator {
+    #[inline(always)]
+    pub async fn next(&mut self) -> HummockResult<()> {
+        match self {
+            Self::Forward(ref mut iter) => iter.next().await,
+            Self::Backward(ref mut iter) => iter.next().await,
+        }
+    }
+
+    #[inline(always)]
+    pub fn key(&self) -> &[u8] {
+        match self {
+            Self::Forward(iter) => iter.key(),
+            Self::Backward(iter) => iter.key(),
+        }
+    }
+
+    #[inline(always)]
+    pub fn value(&self) -> &[u8] {
+        match self {
+            Self::Forward(iter) => iter.value(),
+            Self::Backward(iter) => iter.value(),
+        }
+    }
+
+    #[inline(always)]
+    pub async fn rewind(&mut self) -> HummockResult<()> {
+        match self {
+            Self::Forward(ref mut iter) => iter.rewind().await,
+            Self::Backward(ref mut iter) => iter.rewind().await,
+        }
+    }
+
+    #[inline(always)]
+    pub async fn seek(&mut self, user_key: &[u8]) -> HummockResult<()> {
+        match self {
+            Self::Forward(ref mut iter) => iter.seek(user_key).await,
+            Self::Backward(ref mut iter) => iter.seek(user_key).await,
+        }
+    }
+
+    #[inline(always)]
+    pub fn is_valid(&self) -> bool {
+        match self {
+            Self::Forward(iter) => iter.is_valid(),
+            Self::Backward(iter) => iter.is_valid(),
+        }
+    }
+
+    pub fn collect_local_statistic(&self, stats: &mut StoreLocalStatistic) {
+        match self {
+            DirectedUserIterator::Forward(iter) => iter.collect_local_statistic(stats),
+            DirectedUserIterator::Backward(iter) => iter.collect_local_statistic(stats),
+        }
+    }
+}
